+++ includes/test_ssh.php
@@ -44,6 +44,7 @@ function ssh_try ($test) {
 
 	// default result
 	$results['result'] = 'ok';
+	$results['curl'] = false;
 	$results['time'] = time();
 	$results['error'] = '';
 	$results['result_search'] = 'not tested';
@@ -178,9 +179,6 @@ function ssh_try ($test) {
 	}
 
 	$t = microtime(true) - $s;
-
-	$results['options']['connect_time'] = $results['options']['total_time'] = $results['options']['namelookup_time'] = round($t, 4);
-
 
 	// If we have set a failed search string, then ignore the normal searches and only alert on it
 	if ($test['search_failed'] != '') {